refac: ai 마크다운 코드 분리 (#28)
